requirements.txt
duckduckgo-search>=6.0.0
transformers>=5.0.0
torch>=2.0.0
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
tqdm>=4.65.0
numpy>=1.24.0
PyYAML>=6.0
